Update patch set 1
Patch Set 1:

It's worth mentioning that the open source camera HAL for n80xx does not have this issue. I'll throw a patch up on Gerrit sometime this weekend for that, just wanted to try to figure out what actually works and doesn't work with that first (been running Geecko's ION patch which breaks camera, so I wanted to try fixing the open source HAL with that first).

Patch-set: 1